Move PathFilters to detekt-utils
