+++ detekt-cli/build.gradle.kts
@@ -22,6 +22,7 @@ val pluginsJarFiles by configurations.resolvable("pluginsJarFiles") {
 dependencies {
     implementation(libs.jcommander)
     implementation(projects.detektTooling)
+    implementation(projects.detektUtils)
     implementation(libs.kotlin.compilerEmbeddable) {
         version {
             strictly(libs.versions.kotlin.get())
